Ports common tests from `Log4j1ConfigurationFactoryTest`
